fix: account update and delete

+++ accounts-service/src/main/java/ru/barkhatnat/service/AccountServiceImpl.java
@@ -4,9 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.barkhatnat.DTO.AccountDto;
+import ru.barkhatnat.DTO.AccountResponseDto;
+import ru.barkhatnat.entity.Operation;
 import ru.barkhatnat.repositories.AccountRepository;
 import ru.barkhatnat.entity.Account;
-import ru.barkhatnat.entity.User;
+import ru.barkhatnat.utils.AccountMapper;
+import ru.barkhatnat.utils.CurrentUser;
 import ru.barkhatnat.utils.SecurityUtil;
 
 import java.math.BigDecimal;
@@ -20,17 +23,26 @@ import java.util.Optional;
 public class AccountServiceImpl implements AccountService {
     private final AccountRepository accountRepository;
     private final UserServiceImpl userService;
+    private final CurrentUser currentUser;
+    private final AccountMapper accountMapper;
 
     @Override
     @Transactional
     public Iterable<Account> findAllAccounts() {
-        return userService.findAllUserAccounts(getUser());
+        Integer id = SecurityUtil.getCurrentUserDetails().getUserId();
+        return userService.findAllUserAccounts(id);
+    }
+
+    @Override
+    public Iterable<Operation> findAllAccountOperations(Account account) {
+        return account.getOperations();
     }
 
     @Override
     @Transactional
-    public Account createAccount(AccountDto accountDto) {
-        return accountRepository.save(new Account(accountDto.title(), accountDto.balance(), getUser(), getCreationDate()));
+    public AccountResponseDto createAccount(AccountDto accountDto) {
+        Account account = accountRepository.save(new Account(accountDto.title(), accountDto.balance(), currentUser.getUser(), getCreationDate()));
+        return accountMapper.toAccountResponseDto(account);
     }
 
     @Override
@@ -42,9 +54,14 @@ public class AccountServiceImpl implements AccountService {
     @Override
     @Transactional
     public void updateAccount(Integer id, String title, BigDecimal balance) {
+
         accountRepository.findById(id).ifPresentOrElse(account -> {
-                    account.setTitle(title);
-                    account.setBalance(balance);
+                    if (account.getUser() != null && account.getUser().getId().equals(SecurityUtil.getCurrentUserDetails().getUserId())) {
+                        account.setTitle(title);
+                        account.setBalance(balance);
+                    } else {
+                        throw new IllegalArgumentException("You are not authorized to modify this account");
+                    }
                 }, () -> {
                     throw new NoSuchElementException();
                 }
@@ -54,19 +71,19 @@ public class AccountServiceImpl implements AccountService {
     @Override
     @Transactional
     public void deleteAccount(int id) {
-        accountRepository.deleteById(id);
+        accountRepository.findById(id).ifPresentOrElse(account -> {
+                    if (account.getUser() != null && account.getUser().getId().equals(SecurityUtil.getCurrentUserDetails().getUserId())) {
+                        accountRepository.deleteById(id);
+                    } else {
+                        throw new IllegalArgumentException("You are not authorized to modify this account");
+                    }
+                }, () -> {
+                    throw new NoSuchElementException();
+                }
+        );
     }
 
     private Timestamp getCreationDate() {
         return Timestamp.from(Instant.now());
-    }
-
-    private User getUser() {
-        Integer id = SecurityUtil.getCurrentUserDetails().getUserId();
-        Optional<User> user = userService.findUser(id);
-        if (user.isEmpty()) {
-            throw new NoSuchElementException(); //TODO сделать кастомный эксепшн
-        }
-        return user.get();
     }
 }
